Remove duplicate media query breakpoint and reuse MUI theme
